requirements.txt
SQLAlchemy~=2.0.36
pydantic~=2.9.2
python-dotenv~=1.0.1
fastapi~=0.115.4
python-jose~=3.3.0